Refactor BeanContextBuilder extract interface, use static newBuilder()
