+++ inject-test/src/test/java/org/example/coffee/FactoryTest.java
@@ -1,7 +1,6 @@
 package org.example.coffee;
 
 import io.avaje.inject.BeanContext;
-import io.avaje.inject.BeanContextBuilder;
 import org.example.coffee.factory.BFact;
 import org.junit.jupiter.api.Test;
 
@@ -12,9 +11,7 @@ public class FactoryTest {
   @Test
   public void test() {
 
-    try (BeanContext context = new BeanContextBuilder()
-      .build()) {
-
+    try (BeanContext context = BeanContext.newBuilder().build()) {
       BFact bean = context.getBean(BFact.class);
       String b = bean.b();
       assertThat(b).isNotNull();
