+++ inject-test/src/test/java/org/example/coffee/circular/CircularDependencyTest.java
@@ -1,7 +1,6 @@
 package org.example.coffee.circular;
 
 import io.avaje.inject.BeanContext;
-import io.avaje.inject.BeanContextBuilder;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -10,8 +9,7 @@ class CircularDependencyTest {
 
   @Test
   void wire() {
-    try (BeanContext context = new BeanContextBuilder()
-      .build()) {
+    try (BeanContext context = BeanContext.newBuilder().build()) {
       assertThat(context.getBean(CircA.class)).isNotNull();
       assertThat(context.getBean(CircB.class)).isNotNull();
       assertThat(context.getBean(CircC.class)).isNotNull();
